requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
build==1.4.2
certifi==2026.2.25
charset-normalizer==3.4.7
chromadb==1.5.5
click==8.3.2
durationpy==0.10
FalkorDB==1.6.0
fastapi==0.135.3
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.3.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.9.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
numpy==2.4.4
oauthlib==3.3.1
onnxruntime==1.24.4
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.8
overrides==7.7.0
packaging==26.0
protobuf==6.33.6
pybase64==1.4.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
zipp==3.23.0
